订单服务接口

+++ fmkj-order/order-server/src/main/java/com/fmkj/order/server/controller/ProductController.java
@@ -4,16 +4,20 @@ import com.baomidou.mybatisplus.plugins.Page;
 import com.fmkj.common.base.BaseApiService;
 import com.fmkj.common.base.BaseController;
 import com.fmkj.common.base.BaseResult;
+import com.fmkj.common.base.BaseResultEnum;
 import com.fmkj.common.constant.LogConstant;
+import com.fmkj.common.util.StringUtils;
 import com.fmkj.order.dao.domain.ProductInfo;
 import com.fmkj.order.server.annotation.OrderLog;
 import com.fmkj.order.server.service.ProductService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.DependsOn;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Date;
 import java.util.HashMap;
 
 /**
@@ -22,7 +26,7 @@ import java.util.HashMap;
 @RestController
 @RequestMapping("/product")
 @DependsOn("springContextHolder")
-@Api(tags ={ "商品服务"},description = "商品服务接口")
+@Api(tags ={ "商品服务"},description = "商品服务接口-网关路径/api-race")
 public class ProductController extends BaseController<ProductInfo, ProductService> implements BaseApiService<ProductInfo> {
 
 
@@ -43,9 +47,12 @@ public class ProductController extends BaseController<ProductInfo, ProductServic
 
     @ApiOperation(value="删除商品", notes="根据ID删除商品")
     @OrderLog(module= LogConstant.HC_PRODUCT, actionDesc = "删除商品")
-    @PostMapping("/deleteById")
-    public BaseResult deleteById(@RequestParam String id){
+    @DeleteMapping("/deleteById")
+    public BaseResult deleteById(String id){
         try {
+            if(id == null){
+                return new BaseResult(BaseResultEnum.BLANK.getStatus(), "ID不能为空", "ID不能为空");
+            }
             return super.deleteById(id);
         } catch (Exception e) {
             throw new RuntimeException("删除异常：" + e.getMessage());
@@ -57,6 +64,7 @@ public class ProductController extends BaseController<ProductInfo, ProductServic
     @PostMapping("/addProduct")
     public BaseResult addProduct(ProductInfo productInfo){
         try {
+            productInfo.setCreateTime(new Date());
             return super.insert(productInfo);
         } catch (Exception e) {
             throw new RuntimeException("新增异常：" + e.getMessage());
@@ -68,6 +76,10 @@ public class ProductController extends BaseController<ProductInfo, ProductServic
     @PostMapping("/updateProduct")
     public BaseResult updateProduct(ProductInfo productInfo){
         try {
+            if(StringUtils.isNull(productInfo) || productInfo.getId() == null){
+                return new BaseResult(BaseResultEnum.BLANK.getStatus(), "ID不能为空", "ID不能为空");
+            }
+            productInfo.setUpdateTime(new Date());
             return super.updateById(productInfo);
         } catch (Exception e) {
             throw new RuntimeException("修改异常：" + e.getMessage());
